Murray\The New Life: Chapters 41-52
proofed; formatted

This is the worst book scan I've seen. Margin notes and the odd
substitution of "o" for "v" made this quite challenging. I half think
that I could have typed this from scratch more easily.

Next step is to add in all the marginal verse references as footnotes!